server/requirements-dev.txt
flask-restful==0.3.5
flask-sqlalchemy==2.1
sqlalchemy-migrate==0.10.0
mysql-python==1.2.5
openpyxl==2.3.5
factory-boy==2.8.1
mock==2.0.0
google-auth==0.10.0
httplib2==0.10.3
uritemplate==3.0.0
google-api-python-client==1.6.2
requests==2.13.0
requests-toolbelt==0.7.1
coverage==4.4.1
